+++ beacon_node/network/src/sync/range_sync/range.rs
@@ -371,12 +371,11 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::router::BeaconProcessorSend;
     use crate::service::RequestId;
     use crate::NetworkMessage;
 
     use super::*;
-    use crate::beacon_processor::WorkEvent as BeaconWorkEvent;
+    use crate::beacon_processor::{WorkEvent as BeaconWorkEvent, BeaconProcessorSend};
     use beacon_chain::builder::Witness;
     use beacon_chain::eth1_chain::CachingEth1Backend;
     use beacon_chain::parking_lot::RwLock;
@@ -604,7 +603,7 @@ mod tests {
         let cx = SyncNetworkContext::new(
             network_tx,
             globals.clone(),
-            beacon_processor_tx,
+            BeaconProcessorSend(beacon_processor_tx),
             log.new(o!("component" => "network_context")),
         );
         let test_rig = TestRig {
